Update Main.java
fixed problem where after you land on an item spot and you go to another spot, the "i" was still there and you could pick up more items if you keep landing on there. Added a reveal and removeCharAtLoc to change spot to "n" after it has been landed on.